chore: move findClosestStart to helpers
